Improve parallelism (#5)
Creates 2 processes per core on the CPU for running the minimum edit distance algo across entire steam game list comparing to a given game title. This is the major bottleneck of the program - I plan to now speed it up by building a map of titles and first trying to find the title in the map.